LPD-25757 move to analyticsSettings
